requirements.txt
discord.py
fastapi
uvicorn
aiohttp
pyjwt
python-multipart
pydantic